save and restore state of timer across reloads
This is because say you're in iOS in app mode. You start the timer, you switch to a music
app or messaging app for a moment. You come back to your timer. Without this
the timer is reset.

A question is should the timer keep running or pause. As is the timer appears to have
been running. In other words if you leave the page for 3 minutes when you come
back 3 minutes will have passed. If you wanted to pause you should have paused
before you switched away? Not sure which is best